tests/wpt/tests/web-animations/responsive/verticalAlign.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<div id='container'>
<div id='element'></div>
</div>

<script>
var element = document.getElementById('element');
var container = document.getElementById('container');

test(function() {
    element.style.fontSize = '10px';
    var player = element.animate([{verticalAlign: '3em'}, {verticalAlign: '5em'}], 10);
    player.pause();
    player.currentTime = 5;
    element.style.fontSize = '20px';
    assert_equals(getComputedStyle(element).verticalAlign, '80px');
}, 'verticalAlign responsive to style changes');

</script>
